+++ changelog.md
@@ -1,6 +1,7 @@
 v1.0.0:
 * Extended the logging support to many components and classes in the KTB with the idea to support better troubleshooting. (Closes #101)
 * Add a filter to delete only topic configurations that are explicitly set, not the ones that use default values. (closes #99)
+* Fix a bug with wrongly set PatternType for Consumer group acls level. (closes #111)
 
 v1.0.0-rc2:
 * Check for required configuration values for the configuration of RBAC, if not present it raises a Configuration error